+++ src/machinable/execution/schedule.py
@@ -3,8 +3,8 @@ import random
 
 from ..config.parser import ModuleClass
 from ..core.component import Component as BaseComponent
-from ..utils.traits import Jsonable
 from ..utils.identifiers import encode_experiment_id, generate_component_id
+from ..utils.traits import Jsonable
 from ..utils.utils import generate_seed