+++ src/machinable/project/project.py
@@ -6,11 +6,6 @@ import os
 import pickle
 import sys
 
-import gitignore_parser
-from fs.zipfs import WriteZipFS
-
-from machinable.utils.utils import is_valid_variable_name
-
 from ..config.loader import from_callable as load_config_from_callable
 from ..config.loader import from_file as load_config_file
 from ..config.loader import from_string as load_config_from_string
@@ -22,6 +17,7 @@ from ..core.settings import get_settings
 from ..registration import Registration
 from ..utils.dicts import update_dict
 from ..utils.traits import Jsonable
+from ..utils.utils import is_valid_variable_name
 from ..utils.vcs import get_commit
 from .manager import fetch_imports
 
@@ -241,6 +237,9 @@ class Project(Jsonable):
         # Returns
         True on completion
         """
+        import gitignore_parser
+        from fs.zipfs import WriteZipFS
+
         if opener is None:
             opener = open