+++ src/machinable/server/graphql/types/query/__init__.py
@@ -1,4 +1,3 @@
 from .query_type import query
-
 # resolvers
 from .server import resolve_server