+++ src/machinable/storage/storage.py
@@ -1,8 +1,8 @@
 import os
 from typing import List, Tuple, Union
 
-from ..utils.identifiers import decode_experiment_id
 from ..filesystem import open_fs
+from ..utils.identifiers import decode_experiment_id
 from .collections import ExperimentCollection
 from .experiment import ExperimentStorage