+++ tests/execution/execution_resolver_test.py
@@ -1,7 +1,7 @@
 import pytest
 
 from machinable import Engine, execute
-from machinable.engines import Detached
+from machinable.engines.detached_engine import DetachedEngine as Detached
 from machinable.utils.importing import resolve_instance