+++ tests/test_project/engines/__init__.py
@@ -1,3 +1,3 @@
-from machinable.engines import Native
+from machinable.engines.native_engine import NativeEngine as Native
 
 Native()